+++ lib/src/models/insmet_state.dart
@@ -1,0 +1,14 @@
+enum InsmetState {
+  OccasionalShowers,
+  ScatteredShowers,
+  IsolatedShowers,
+  AfternoonShowers,
+  RainShowers,
+  PartlyCloudy,
+  Cloudy,
+  Sunny,
+  Storms,
+  AfternoonStorms,
+  MorningScatteredShowers,
+  Winds,
+}