+++ lib/src/models/weather_model.dart
@@ -1,19 +1,23 @@
-import 'package:cuba_weather_insmet_dart/cuba_weather_insmet_dart.dart';
-
 import 'package:cuba_weather_dart/src/models/models.dart';
 
 class WeatherModel {
-  final String cityName;
-  final int temperature;
-  final int temperatureMax;
-  final int temperatureMin;
-  final DateTime dateTime;
-  final double pressure;
-  final double humidity;
-  final double windVelocity;
-  final String windDirection;
-  final State state;
-  final forecasts = List<WeatherForecastModel>();
+  String cityName;
+  int temperature;
+  int temperatureMax;
+  int temperatureMin;
+  DateTime dateTime;
+  double pressure;
+  double humidity;
+  double windVelocity;
+  CardinalPoint windDirection;
+  String windDirectionDescription;
+  int windDirectionDegree;
+  double windDirectionRadians;
+  String weatherForecast;
+  String droughtStatus;
+  InsmetState state;
+  String stateDescription;
+  var forecasts = List<WeatherForecastModel>();
 
   WeatherModel({
     this.cityName,
@@ -25,20 +29,27 @@ class WeatherModel {
     this.humidity,
     this.windVelocity,
     this.windDirection,
+    this.windDirectionDescription,
+    this.windDirectionDegree,
+    this.windDirectionRadians,
+    this.weatherForecast,
+    this.droughtStatus,
     this.state,
+    this.stateDescription,
   });
 
-  void addForecast({
-    int day,
-    int temperatureMax,
-    int temperatureMin,
-    State state,
-  }) {
+  void addForecast(
+      {int day,
+      int temperatureMax,
+      int temperatureMin,
+      InsmetState state,
+      String stateDescription}) {
     forecasts.add(WeatherForecastModel(
       day: day,
       temperatureMax: temperatureMax,
       temperatureMin: temperatureMin,
       state: state,
+      stateDescription: stateDescription,
     ));
   }
 
@@ -54,7 +65,12 @@ class WeatherModel {
     result.write('Pressure: $pressure hPa\n');
     result.write('Wind Velocity: $windVelocity Km/h\n');
     result.write('Wind Direction: $windDirection\n');
+    result.write('Wind Direction Degree: ${windDirectionDegree}\n');
+    result.write('Wind Direction Radians: ${windDirectionRadians}\n');
+    result.write('Today\'s Weather Forecast: $weatherForecast\n');
+    result.write('Drought Status: $droughtStatus\n');
     result.write('State: $state\n');
+    result.write('State Description: $stateDescription\n');
     for (var i = 0; i < forecasts.length; ++i) {
       result.write('${forecasts[i]}');
       if (i + 1 < forecasts.length) {