+++ lib/src/repositories/repositories.dart
@@ -1,1 +1,7 @@
-export 'weather_repository.dart';
+export 'insmet_forecast_repository.dart';
+export 'insmet_marine_forecast_repository.dart';
+export 'insmet_source_repository.dart';
+export 'insmet_weather_repository.dart';
+export 'municipality_repository.dart';
+export 'redcuba_source_repository.dart';
+export 'redcuba_weather_repository.dart';